% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_enlista.R
\name{fn_enlista}
\alias{fn_enlista}
\title{fn_enlista}
\usage{
fn_enlista(x)
}
\arguments{
\item{x}{A vector that contains multiple elements.}
}
\value{
A vector that inserts a comma after each element, adds
 'y' to the last element, and if the last element is a word
 starting with 'i', it instead adds 'e'.
}
\description{
This function is included in the Redactum package.
}
\examples{
 \dontrun{
 # The last word does not begin with "I".
 x <- c("Pedro", "Juan", "Mario", "Emilio", "Iker", "Hugo")
 xx <- fn_enlista(x)

 # Combine it with paste0().
 paste0("Los amigos de Armando son ", xx,".")

 # The last word begins  with "I".
 x <- c("Pedro", "Juan", "Mario", "Emilio", "Iker")
 xx <- fn_enlista(x)

 # The last word begins  with "i".
 x <- c("península", "cordillera", "isla")
 xx <- fn_enlista(x)
 }

}
