% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_cient.R
\name{n_cient}
\alias{n_cient}
\title{n_cient}
\usage{
n_cient(name)
}
\arguments{
\item{name}{A character string containing the scientific name.}
}
\value{
A character string containing the abbreviated scientific name.
}
\description{
Abbreviated scientific name
}
\details{
The abbreviated name is derived from the scientific name. To create the abbreviation, the first letter of the genus is taken, written in uppercase, and followed by a period. The second part of the scientific name is then added after a space.
}
\examples{

n_cient("Homo sapiens")

name <- "Hydrolagus melanophasma"
n_cient(name)

}
