% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnum.R
\name{nnum}
\alias{nnum}
\title{nnum}
\usage{
nnum(x, nnames)
}
\arguments{
\item{x}{an integer number.}

\item{nnames}{a character vector that contains the names of numbers.}
}
\value{
It returns the name of a number in the language specified by the user.
}
\description{
Returns the name of a specified number.
}
\details{
This function returns the name of a given number, specifically for numbers ranging from 0 to 10.
It uses an array that defines the names of the numbers from one to ten, including zero at the end.
The user specifies the names of the numbers in the chosen language.
}
\examples{
\dontrun{
## Number in English.
nnames <- c("one", "two", "three",
 "four", "five", "six",
  "seven", "eight", "nine",
   "ten", "zero")

x <- 2
nnum(x,nnames)

## Number in Spanish.
nnames <- c("uno", "dos", "tres",
 "cuatro", "cinco", "seis",
 "siete", "ocho", "nueve",
  "diez","cero")

x <- 2
nnum(x,nnames)
  }

}
