% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_cient.R
\name{n_cient}
\alias{n_cient}
\title{n_cient}
\usage{
n_cient(name)
}
\arguments{
\item{name}{A character string containing the scientific name.}
}
\value{
A character string containing the abbreviated scientific name.
}
\description{
Abbreviated scientific name
}
\details{
The abbreviated name is derived from the scientific name. To create the abbreviation, the first letter of the genus is taken, written in uppercase, and followed by a period. The specific name is added after a space in the scientific name.

Within the function, a condition is established. If the scientific name refers to a species that is not specified under a particular genus, such as Hydrolagus sp., or to multiple species within a genus, like Hydrolagus spp., the name will not be abbreviated. However, if the scientific name designates a specific species, it will be modified; for instance, Hydrolagus melanophasma will be abbreviated to H. melanophasma.

The function operates correctly with or without periods in abbreviations, such as "sp" or "sp."
}
\examples{

n_cient("Homo sapiens")

name <- "Hydrolagus melanophasma"
n_cient(name)

## Example: Using the abbreviation "sp" without a period.
name <- "Hydrolagus sp"
n_cient(name)

## Example: Using the abbreviation "spp." with a period.
name <- "Hydrolagus spp."
n_cient(name)

}
